Enable a bunch of remote tests
