+++ test/e2e/generate_kube_test.go
@@ -1,5 +1,3 @@
-// +build !remote
-
 package integration
 
 import (
