+++ test/e2e/libpod_suite_test.go
@@ -9,30 +9,10 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/containers/libpod/v2/pkg/cgroups"
 	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 )
 
 func SkipIfRemote() {
-}
-
-func SkipIfCgroupV1() {
-	cgroupsv2, err := cgroups.IsCgroup2UnifiedMode()
-	Expect(err).To(BeNil())
-
-	if !cgroupsv2 {
-		Skip("Skip on systems with cgroup V1 systems")
-	}
-}
-
-func SkipIfCgroupV2() {
-	cgroupsv2, err := cgroups.IsCgroup2UnifiedMode()
-	Expect(err).To(BeNil())
-
-	if cgroupsv2 {
-		Skip("Skip on systems with cgroup V2 systems")
-	}
 }
 
 func SkipIfRootless() {
