+++ test/e2e/pull_test.go
@@ -1,5 +1,3 @@
-// +build !remote
-
 package integration
 
 import (
@@ -237,6 +235,7 @@ var _ = Describe("Podman pull", func() {
 	})
 
 	It("podman pull from docker-archive", func() {
+		SkipIfRemote()
 		podmanTest.RestoreArtifact(ALPINE)
 		tarfn := filepath.Join(podmanTest.TempDir, "alp.tar")
 		session := podmanTest.PodmanNoCache([]string{"save", "-o", tarfn, "alpine"})
@@ -255,6 +254,7 @@ var _ = Describe("Podman pull", func() {
 	})
 
 	It("podman pull from oci-archive", func() {
+		SkipIfRemote()
 		podmanTest.RestoreArtifact(ALPINE)
 		tarfn := filepath.Join(podmanTest.TempDir, "oci-alp.tar")
 		session := podmanTest.PodmanNoCache([]string{"save", "--format", "oci-archive", "-o", tarfn, "alpine"})
@@ -273,6 +273,7 @@ var _ = Describe("Podman pull", func() {
 	})
 
 	It("podman pull from local directory", func() {
+		SkipIfRemote()
 		podmanTest.RestoreArtifact(ALPINE)
 		dirpath := filepath.Join(podmanTest.TempDir, "alpine")
 		os.MkdirAll(dirpath, os.ModePerm)
@@ -297,6 +298,7 @@ var _ = Describe("Podman pull", func() {
 	})
 
 	It("podman pull from local OCI directory", func() {
+		SkipIfRemote()
 		podmanTest.RestoreArtifact(ALPINE)
 		dirpath := filepath.Join(podmanTest.TempDir, "alpine")
 		os.MkdirAll(dirpath, os.ModePerm)
