+++ test/e2e/run_passwd_test.go
@@ -1,5 +1,3 @@
-// +build !remote
-
 package integration
 
 import (
